feat : 결제, 환불, 주문 사이 안 맞는 연관 코드 수정 - 결제 관련 버그도 수정.

+++ src/main/java/org/example/murderhelp/domain/payment/dto/PaymentConfirmResponse.java
@@ -5,7 +5,7 @@ import org.example.murderhelp.domain.payment.entity.Payment;
 public record PaymentConfirmResponse(
         Long paymentId,
         Long orderId,
-        int amount,
+        Long amount,
         String paymentStatus,
         String orderStatus,
         String message
